requirements.txt
python-dotenv==1.0.1
jqdatasdk==1.9.7
SQLAlchemy==2.0.37
psycopg2==2.9.10
tqdm==4.67.1
